requirements.txt
uvicorn[standard]
gunicorn
fastapi[all]
sqlmodel
asyncpg
python-dotenv
alembic
sqlalchemy[asyncio]
supabase
psycopg2-binary
asyncpg